+++ lib/wasm-interface/src/validate.rs
@@ -1,16 +1,16 @@
-//! Validate a wasm module given a contract.
+//! Validate a wasm module given a interface.
 //!
-//! This checks that all imports are specified in the contract and that their types
-//! are correct, as well as that all exports that the contract expects are exported
+//! This checks that all imports are specified in the interface and that their types
+//! are correct, as well as that all exports that the interface expects are exported
 //! by the module and that their types are correct.
 
-use crate::{Contract, Export, Import, WasmType};
+use crate::{Interface, Export, Import, WasmType};
 use std::collections::HashMap;
 use wasmparser::{ExternalKind, FuncType, GlobalType, ImportSectionEntryType};
 
 pub fn validate_wasm_and_report_errors(
     wasm: &[u8],
-    contract: &Contract,
+    interface: &Interface,
 ) -> Result<(), WasmValidationError> {
     use wasmparser::WasmDecoder;
 
@@ -50,7 +50,7 @@ pub fn validate_wasm_and_report_errors(
                                 ),
                             }
                         })?;
-                    if let Some(val) = contract.imports.get(&Import::format_key(module, field)) {
+                    if let Some(val) = interface.imports.get(&Import::format_key(module, field)) {
                         if let Import::Global { var_type, .. } = val {
                             if *var_type != global_type {
                                 errors.push(format!(
@@ -66,7 +66,7 @@ pub fn validate_wasm_and_report_errors(
                         }
                     } else {
                         errors.push(format!(
-                            "Global import \"{}\" not found in the specified contract",
+                            "Global import \"{}\" not found in the specified interface",
                             field
                         ));
                     }
@@ -96,34 +96,34 @@ pub fn validate_wasm_and_report_errors(
         }
     }
 
-    validate_imports(&import_fns, &type_defs, contract, &mut errors);
-    validate_export_fns(&export_fns, &type_defs, contract, &mut errors);
-    validate_export_globals(&export_globals, &global_types, contract, &mut errors);
+    validate_imports(&import_fns, &type_defs, interface, &mut errors);
+    validate_export_fns(&export_fns, &type_defs, interface, &mut errors);
+    validate_export_globals(&export_globals, &global_types, interface, &mut errors);
 
     if errors.is_empty() {
         Ok(())
     } else {
-        Err(WasmValidationError::ContractViolated { errors: errors })
+        Err(WasmValidationError::InterfaceViolated { errors: errors })
     }
 }
 
 /// Validates the import functions, checking the name and type against the given
-/// `Contract`
+/// `Interface`
 fn validate_imports(
     import_fns: &HashMap<(String, String), u32>,
     type_defs: &Vec<FuncType>,
-    contract: &Contract,
+    interface: &Interface,
     errors: &mut Vec<String>,
 ) {
     for (key, val) in import_fns.iter() {
-        if let Some(contract_def) = contract.imports.get(key) {
+        if let Some(interface_def) = interface.imports.get(key) {
             let type_sig = if let Some(v) = type_defs.get(*val as usize) {
                 v
             } else {
                 errors.push(format!("Invalid type reference {}", val));
                 continue;
             };
-            if let Import::Func { params, result, .. } = contract_def {
+            if let Import::Func { params, result, .. } = interface_def {
                 debug_assert!(type_sig.form == wasmparser::Type::Func);
                 for (i, param) in type_sig
                     .params
@@ -135,7 +135,7 @@ fn validate_imports(
                     match param {
                         Ok(t) => {
                             if params.get(i).is_none() {
-                                errors.push(format!("Found {} args but the contract only expects {} for imported function \"{}\" \"{}\"", i, params.len(), &key.0, &key.1));
+                                errors.push(format!("Found {} args but the interface only expects {} for imported function \"{}\" \"{}\"", i, params.len(), &key.0, &key.1));
                                 continue;
                             }
                             if t != params[i] {
@@ -161,7 +161,7 @@ fn validate_imports(
                     match ret {
                         Ok(t) => {
                             if result.get(i).is_none() {
-                                errors.push(format!("Found {} returns but the contract only expects {} for imported function \"{}\" \"{}\"", i, params.len(), &key.0, &key.1));
+                                errors.push(format!("Found {} returns but the interface only expects {} for imported function \"{}\" \"{}\"", i, params.len(), &key.0, &key.1));
                                 continue;
                             }
 
@@ -180,7 +180,7 @@ fn validate_imports(
                 }
             }
         } else {
-            // we didn't find the import at all in the contract
+            // we didn't find the import at all in the interface
             // TODO: improve error messages by including type information
             errors.push(format!("Missing import \"{}\" \"{}\"", key.0, key.1));
         }
@@ -188,22 +188,22 @@ fn validate_imports(
 }
 
 /// Validates the export functions, checking the name and type against the given
-/// `Contract`
+/// `Interface`
 fn validate_export_fns(
     export_fns: &HashMap<String, u32>,
     type_defs: &Vec<FuncType>,
-    contract: &Contract,
+    interface: &Interface,
     errors: &mut Vec<String>,
 ) {
     for (key, val) in export_fns.iter() {
-        if let Some(contract_def) = contract.exports.get(key) {
+        if let Some(interface_def) = interface.exports.get(key) {
             let type_sig = if let Some(v) = type_defs.get(*val as usize) {
                 v
             } else {
                 errors.push(format!("Invalid type reference {}", val));
                 continue;
             };
-            if let Export::Func { params, result, .. } = contract_def {
+            if let Export::Func { params, result, .. } = interface_def {
                 debug_assert!(type_sig.form == wasmparser::Type::Func);
                 for (i, param) in type_sig
                     .params
@@ -215,7 +215,7 @@ fn validate_export_fns(
                     match param {
                         Ok(t) => {
                             if params.get(i).is_none() {
-                                errors.push(format!("Found {} args but the contract only expects {} for exported function \"{}\"", i, params.len(), &key));
+                                errors.push(format!("Found {} args but the interface only expects {} for exported function \"{}\"", i, params.len(), &key));
                                 continue;
                             }
                             if t != params[i] {
@@ -236,7 +236,7 @@ fn validate_export_fns(
                     match ret {
                         Ok(t) => {
                             if result.get(i).is_none() {
-                                errors.push(format!("Found {} returns but the contract only expects {} for exported function \"{}\"", i, params.len(), &key));
+                                errors.push(format!("Found {} returns but the interface only expects {} for exported function \"{}\"", i, params.len(), &key));
                                 continue;
                             }
 
@@ -254,16 +254,16 @@ fn validate_export_fns(
 }
 
 /// Validates the export globals, checking the name and type against the given
-/// `Contract`
+/// `Interface`
 fn validate_export_globals(
     export_globals: &HashMap<String, u32>,
     global_types: &Vec<GlobalType>,
-    contract: &Contract,
+    interface: &Interface,
     errors: &mut Vec<String>,
 ) {
     for (key, val) in export_globals.iter() {
-        if let Some(contract_def) = contract.exports.get(key) {
-            if let Export::Global { var_type, .. } = contract_def {
+        if let Some(interface_def) = interface.exports.get(key) {
+            if let Export::Global { var_type, .. } = interface_def {
                 if global_types.get(*val as usize).is_none() {
                     errors.push(format!(
                         "Invalid wasm, expected {} global types, found {}",
@@ -287,11 +287,11 @@ fn validate_export_globals(
     }
 }
 
-/// Converts Wasmparser's type enum into wasm-contract's type enum
+/// Converts Wasmparser's type enum into wasm-interface's type enum
 /// wasmparser's enum contains things which are invalid in many situations
 ///
 /// Additionally wasmerparser containers more advanced types like references that
-/// wasm-contract does not yet support
+/// wasm-interface does not yet support
 fn wasmparser_type_into_wasm_type(ty: wasmparser::Type) -> Result<WasmType, String> {
     use wasmparser::Type;
     Ok(match ty {
@@ -319,22 +319,22 @@ mod validation_tests {
 )"#;
         let wasm = wabt::wat2wasm(WAT).unwrap();
 
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_import (func "env" "do_panic" (param i32 i64)))
 (assert_import (global "env" "length" (type i32)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(result.is_ok());
 
         // Now set the global import type to mismatch the wasm
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_import (func "env" "do_panic" (param i32 i64)))
 (assert_import (global "env" "length" (type i64)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(
             result.is_err(),
@@ -342,29 +342,29 @@ mod validation_tests {
         );
 
         // Now set the function import type to mismatch the wasm
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_import (func "env" "do_panic" (param i64)))
 (assert_import (global "env" "length" (type i32)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(
             result.is_err(),
             "function import type mismatch causes an error"
         );
 
-        // Now try with a module that has an import that the contract doesn't have
-        let contract_src = r#"
+        // Now try with a module that has an import that the interface doesn't have
+        let interface_src = r#"
 (assert_import (func "env" "do_panic" (param i64)))
 (assert_import (global "env" "length_plus_plus" (type i32)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(
             result.is_err(),
-            "all imports must be covered by the contract"
+            "all imports must be covered by the interface"
         );
     }
 
@@ -377,22 +377,22 @@ mod validation_tests {
 )"#;
         let wasm = wabt::wat2wasm(WAT).unwrap();
 
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_export (func "as-set_local-first" (param i32) (result i32)))
 (assert_export (global "num_tries" (type i64)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(result.is_ok());
 
         // Now set the global export type to mismatch the wasm
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_export (func "as-set_local-first" (param i32) (result i32)))
 (assert_export (global "num_tries" (type f32)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(
             result.is_err(),
@@ -400,25 +400,25 @@ mod validation_tests {
         );
 
         // Now set the function export type to mismatch the wasm
-        let contract_src = r#"
+        let interface_src = r#"
 (assert_export (func "as-set_local-first" (param i64) (result i64)))
 (assert_export (global "num_tries" (type i64)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(
             result.is_err(),
             "function export type mismatch causes an error"
         );
 
-        // Now try a contract that requires an export that the module doesn't have
-        let contract_src = r#"
+        // Now try a interface that requires an export that the module doesn't have
+        let interface_src = r#"
 (assert_export (func "as-set_local-first" (param i64) (result i64)))
 (assert_export (global "numb_trees" (type i64)))"#;
-        let contract = parser::parse_contract(contract_src).unwrap();
+        let interface = parser::parse_interface(interface_src).unwrap();
 
-        let result = validate_wasm_and_report_errors(&wasm[..], &contract);
+        let result = validate_wasm_and_report_errors(&wasm[..], &interface);
 
         assert!(result.is_err(), "missing a required export is an error");
     }
@@ -427,6 +427,6 @@ mod validation_tests {
 #[derive(Debug)]
 pub enum WasmValidationError {
     InvalidWasm { error: String },
-    ContractViolated { errors: Vec<String> },
+    InterfaceViolated { errors: Vec<String> },
     UnsupportedType { error: String },
 }