+++ src/interfaces.rs
@@ -3,49 +3,49 @@ use crate::sql;
 
 use rusqlite::{params, Connection, TransactionBehavior};
 
-pub fn contract_exists(
+pub fn interface_exists(
     conn: &mut Connection,
-    contract_name: &str,
+    interface_name: &str,
     version: &str,
 ) -> Result<bool, failure::Error> {
     let mut stmt = conn.prepare(sql::WASM_CONTRACT_EXISTENCE_CHECK)?;
-    Ok(stmt.exists(params![contract_name, version])?)
+    Ok(stmt.exists(params![interface_name, version])?)
 }
 
-pub fn load_contract_from_db(
+pub fn load_interface_from_db(
     conn: &mut Connection,
-    contract_name: &str,
+    interface_name: &str,
     version: &str,
-) -> Result<wasm_contract::Contract, failure::Error> {
+) -> Result<wasm_interface::Interface, failure::Error> {
     let mut stmt = conn.prepare(sql::GET_WASM_CONTRACT)?;
-    let contract_string: String =
-        stmt.query_row(params![contract_name, version], |row| Ok(row.get(0)?))?;
+    let interface_string: String =
+        stmt.query_row(params![interface_name, version], |row| Ok(row.get(0)?))?;
 
-    wasm_contract::parser::parse_contract(&contract_string).map_err(|e| {
+    wasm_interface::parser::parse_interface(&interface_string).map_err(|e| {
         format_err!(
-            "Failed to parse contract {} version {} in database: {}",
-            contract_name,
+            "Failed to parse interface {} version {} in database: {}",
+            interface_name,
             version,
             e
         )
     })
 }
 
-pub fn import_contract(
+pub fn import_interface(
     conn: &mut Connection,
-    contract_name: &str,
+    interface_name: &str,
     version: &str,
     content: &str,
 ) -> Result<(), failure::Error> {
-    // fail if we already have this contract
+    // fail if we already have this interface
     {
         let mut key_check = conn.prepare(sql::WASM_CONTRACT_EXISTENCE_CHECK)?;
-        let result = key_check.exists(params![contract_name, version])?;
+        let result = key_check.exists(params![interface_name, version])?;
 
         if result {
             return Err(format_err!(
-                "Contract {}, version {} already exists",
-                contract_name,
+                "Interface {}, version {} already exists",
+                interface_name,
                 version
             ));
         }
@@ -54,10 +54,10 @@ pub fn import_contract(
     let tx = conn.transaction_with_behavior(TransactionBehavior::Immediate)?;
     let time_string = get_current_time_in_format().expect("Could not get current time");
 
-    debug!("Adding contract {:?} {:?}", contract_name, version);
+    debug!("Adding interface {:?} {:?}", interface_name, version);
     tx.execute(
         sql::INSERT_WASM_CONTRACT,
-        params![contract_name, version, time_string, content],
+        params![interface_name, version, time_string, content],
     )?;
 
     tx.commit()?;