+++ src/sql/queries/insert_contract.sql
@@ -1,3 +1,0 @@
-INSERT INTO wasm_contracts
-(contract_name, version, date_added, content)
-VALUES (?1, ?2, ?3, ?4)